// frontend/src/pages/Apply/steps/personal/types.ts
export type PersonalDetailsValues = {
  fullName: string;
  gender: string;
  dateOfBirth: string;
  countryCode: string;
  phoneNumber: string;
  nationality: string;
  countryOfResidence: string;
  residentialAddress: string;
  stateOrProvince: string;
  passportNumber: string;
  passportExpiryDate: string;
};